Move get councillor into bio component
