+++ src/app/councillors/[contactSlug]/components/CouncillorBio.tsx
@@ -1,44 +1,42 @@
-'use client';
-import { useState } from 'react';
-import Image from 'next/image';
-import { CircleUserRound } from 'lucide-react';
+'use server';
+import { createDB } from '@/database/kyselyDb';
+import { Kysely } from 'kysely';
+import { DB } from '@/database/allDbTypes';
+import { Avatar } from '@/components/Avatar';
 
-type CouncillorBioInfo = {
-  photoUrl: string | null;
-  contactName: string;
-  wardName: string;
-  wardId: string;
-  email: string;
-  phone: string | null;
-};
+async function getCouncillor(db: Kysely<DB>, contactSlug: string) {
+  return await db
+    .selectFrom('Councillors')
+    .innerJoin('Contacts', (eb) =>
+      eb.onRef('Contacts.contactSlug', '=', 'Councillors.contactSlug'),
+    )
+    .innerJoin('Wards', (eb) =>
+      eb.onRef('Wards.wardSlug', '=', 'Councillors.wardSlug'),
+    )
+    .select([
+      'Councillors.contactSlug',
+      'Contacts.contactName',
+      'Contacts.email',
+      'Contacts.phone',
+      'Contacts.photoUrl',
+      'Wards.wardName',
+      'Wards.wardId',
+    ])
+    .where('Councillors.contactSlug', '=', contactSlug)
+    .executeTakeFirstOrThrow();
+}
 
-export default function CouncillorBio({
-  councillor,
+export default async function CouncillorBio({
+  contactSlug,
 }: {
-  councillor: CouncillorBioInfo;
+  contactSlug: string;
 }) {
-  const [showFallbackAvatar, setShowFallbackAvatar] = useState(
-    !councillor.photoUrl,
-  );
+  const councillor = await getCouncillor(createDB(), contactSlug);
 
   return (
     <section className="m-8">
       <div className="flex flex-col md:flex-row gap-6 items-center">
-        {showFallbackAvatar ? (
-          <div className="w-48 h-48 light:bg-gray-200 rounded-full flex items-center justify-center">
-            <CircleUserRound size={190} />
-          </div>
-        ) : (
-          <div className="min-w-48 max-w-48 h-48 relative rounded-full overflow-hidden">
-            <Image
-              src={councillor.photoUrl!}
-              alt={councillor.contactName}
-              fill
-              className="object-cover object-top"
-              onError={() => setShowFallbackAvatar(true)}
-            />
-          </div>
-        )}
+        <Avatar src={councillor.photoUrl} size={190} />
 
         <div>
           <h1 className="text-3xl font-bold mb-2">{councillor.contactName}</h1>
