+++ src/app/councillors/[contactSlug]/page.tsx
@@ -16,28 +16,6 @@ export async function generateStaticParams(): Promise<ParamsType[]> {
     .execute();
 }
 
-async function getCouncillor(db: Kysely<DB>, contactSlug: string) {
-  return await db
-    .selectFrom('Councillors')
-    .innerJoin('Contacts', (eb) =>
-      eb.onRef('Contacts.contactSlug', '=', 'Councillors.contactSlug'),
-    )
-    .innerJoin('Wards', (eb) =>
-      eb.onRef('Wards.wardSlug', '=', 'Councillors.wardSlug'),
-    )
-    .select([
-      'Councillors.contactSlug',
-      'Contacts.contactName',
-      'Contacts.email',
-      'Contacts.phone',
-      'Contacts.photoUrl',
-      'Wards.wardName',
-      'Wards.wardId',
-    ])
-    .where('Councillors.contactSlug', '=', contactSlug)
-    .executeTakeFirstOrThrow();
-}
-
 async function getVotesByAgendaItemsForContact(
   db: Kysely<DB>,
   contactSlug: string,
@@ -133,12 +111,12 @@ export default async function CouncillorVotePage(props: {
 }) {
   const { contactSlug } = await props.params;
   const db = createDB();
-  const councillor = await getCouncillor(db, contactSlug);
   const agendaItems = await getVotesByAgendaItemsForContact(db, contactSlug);
 
   return (
     <main className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8">
-      <CouncillorBio councillor={councillor} />
+      <CouncillorBio contactSlug={contactSlug} />
+
       <CouncillorVoteContent agendaItems={agendaItems} />
     </main>
   );
